Add Stable Diffusion client
