+++ backend/tests/test_root.py
@@ -38,7 +38,7 @@ def test_text_endpoint(monkeypatch):
         return mock_resp
 
     monkeypatch.setattr("backend.app.main.requests.post", fake_post)
-    resp = client.post("/text", json={"prompt": "hi"})
+    resp = client.post("/txt", json={"context": "hi"})
     assert resp.status_code == 200
     assert resp.json()["response"] == "ok"
 
@@ -49,8 +49,8 @@ def test_image_endpoint(monkeypatch):
         return {"image": "imgdata"}
 
     monkeypatch.setattr(
-        "backend.app.main.ollama_generate_image", fake_generate
+        "backend.app.main.stablediffusion_generate_image", fake_generate
     )
-    resp = client.post("/image", json={"prompt": "draw"})
+    resp = client.post("/img", json={"prompt": "draw"})
     assert resp.status_code == 200
     assert resp.json()["image"] == "imgdata"
